+++ src/dve/core_engine/message.py
@@ -13,7 +13,7 @@ from typing import Any, ClassVar, Optional, Union
 from pydantic import BaseModel, ValidationError, validator
 from pydantic.dataclasses import dataclass
 
-from dve.core_engine.constants import CONTRACT_ERROR_VALUE_FIELD_NAME, ROWID_COLUMN_NAME
+from dve.core_engine.constants import CONTRACT_ERROR_VALUE_FIELD_NAME, RECORD_INDEX_COLUMN_NAME
 from dve.core_engine.templating import template_object
 from dve.core_engine.type_hints import (
     EntityName,
@@ -116,6 +116,8 @@ class UserMessage:
     "The offending values"
     Category: ErrorCategory
     "The category of error"
+    RecordIndex: Optional[int] = None
+    "The record index where the error occurred (if applicable)"
 
     @property
     def is_informational(self) -> bool:
@@ -187,6 +189,7 @@ class FeedbackMessage:  # pylint: disable=too-many-instance-attributes
         "ErrorMessage",
         "ErrorCode",
         "ReportingField",
+        "RecordIndex",
         "Value",
         "Category",
     ]
@@ -223,15 +226,6 @@ class FeedbackMessage:  # pylint: disable=too-many-instance-attributes
             return str(value[0])
 
         return str(value)
-
-    @validator("record")
-    def _strip_rowid(  # pylint: disable=no-self-argument
-        cls, value: Optional[dict[str, Any]]
-    ) -> Optional[dict[str, Any]]:
-        """Strip the row ID column from the record, if present."""
-        if isinstance(value, dict):
-            value.pop(ROWID_COLUMN_NAME, None)
-        return value
 
     @property
     def is_critical(self) -> bool:
@@ -333,6 +327,7 @@ class FeedbackMessage:  # pylint: disable=too-many-instance-attributes
             error_message,
             self.error_code,
             self.reporting_field_name or reporting_field,
+            (self.record.get(RECORD_INDEX_COLUMN_NAME) if self.record else None),
             value,
             self.category,
         )