+++ src/dve/metadata_parser/domain_types.py
@@ -173,33 +173,67 @@ def permissive_nhs_number(warn_on_test_numbers: bool = False):
     return type("NHSNumber", (NHSNumber, *NHSNumber.__bases__), dict_)
 
 
-# TODO: Make the spacing configurable. Not all downstream consumers want a single space
 class Postcode(types.ConstrainedStr):
     """Postcode constrained string"""
 
     regex: re.Pattern = POSTCODE_REGEX
     strip_whitespace = True
+    apply_normalize = True
 
     @staticmethod
-    def normalize(postcode: str) -> Optional[str]:
+    def normalize(_postcode: str) -> Optional[str]:
         """Strips internal and external spaces"""
-        postcode = postcode.replace(" ", "")
-        if not postcode or postcode.lower() in NULL_POSTCODES:
+        _postcode = _postcode.replace(" ", "")
+        if not _postcode or _postcode.lower() in NULL_POSTCODES:
             return None
-        postcode = postcode.replace(" ", "")
-        return " ".join((postcode[0:-3], postcode[-3:])).upper()
+        _postcode = _postcode.replace(" ", "")
+        return " ".join((_postcode[0:-3], _postcode[-3:])).upper()
 
     @classmethod
     def validate(cls, value: str) -> Optional[str]:  # type: ignore
         """Validates the given postcode"""
-        stripped = cls.normalize(value)
-        if not stripped:
+        if cls.apply_normalize and value:
+            value = cls.normalize(value)  # type: ignore
+
+        if not value:
             return None
 
-        if not cls.regex.match(stripped):
+        if not cls.regex.match(value):
             raise ValueError("Invalid Postcode submitted")
 
-        return stripped
+        return value
+
+
+@lru_cache()
+@validate_arguments
+def postcode(
+    # pylint: disable=R0913
+    strip_whitespace: Optional[bool] = True,
+    to_upper: Optional[bool] = False,
+    to_lower: Optional[bool] = False,
+    strict: Optional[bool] = False,
+    min_length: Optional[int] = None,
+    max_length: Optional[int] = None,
+    curtail_length: Optional[int] = None,
+    regex: Optional[str] = POSTCODE_REGEX,  # type: ignore
+    apply_normalize: Optional[bool] = True,
+) -> type[Postcode]:
+    """Return a formatted date class with a set date format
+    and timezone treatment.
+
+    """
+    dict_ = Postcode.__dict__.copy()
+    dict_["strip_whitespace"] = strip_whitespace
+    dict_["to_upper"] = to_upper
+    dict_["to_lower"] = to_lower
+    dict_["strict"] = strict
+    dict_["min_length"] = min_length
+    dict_["max_length"] = max_length
+    dict_["curtail_length"] = curtail_length
+    dict_["regex"] = regex
+    dict_["apply_normalize"] = apply_normalize
+
+    return type("Postcode", (Postcode, *Postcode.__bases__), dict_)
 
 
 class OrgID(_SimpleRegexValidator):
@@ -226,6 +260,8 @@ class ConFormattedDate(dt.date):
 
     DATE_FORMAT: ClassVar[Optional[str]] = None
     """The specific format of the date as a Python 'strptime' string."""
+    strict: ClassVar[Optional[bool]] = False
+    """Add additional check to ensure that date supplied meets the date format exactly."""
     ge: ClassVar[Optional[dt.date]] = None
     """The earliest date allowed."""
     le: ClassVar[Optional[dt.date]] = None
@@ -246,6 +282,8 @@ class ConFormattedDate(dt.date):
         elif cls.DATE_FORMAT is not None:
             try:
                 date = dt.datetime.strptime(value, cls.DATE_FORMAT).date()
+                if cls.strict and (date.strftime(cls.DATE_FORMAT) != value):
+                    raise ValueError
             except ValueError as err:
                 raise ValueError(
                     f"Unable to parse provided datetime in format {cls.DATE_FORMAT}"
@@ -283,6 +321,7 @@ class ConFormattedDate(dt.date):
 @validate_arguments
 def conformatteddate(
     date_format: Optional[str] = None,
+    strict: Optional[bool] = False,
     ge: Optional[dt.date] = None,  # pylint: disable=invalid-name
     le: Optional[dt.date] = None,  # pylint: disable=invalid-name
     gt: Optional[dt.date] = None,  # pylint: disable=invalid-name
@@ -297,6 +336,7 @@ def conformatteddate(
 
     dict_ = ConFormattedDate.__dict__.copy()
     dict_["DATE_FORMAT"] = date_format
+    dict_["strict"] = strict
     dict_["ge"] = ge
     dict_["le"] = le
     dict_["gt"] = gt
@@ -481,6 +521,11 @@ class FormattedTime(dt.time):
             raise ValueError("Provided time missing timezone, but this is required for this field")
 
         return new_time
+
+    @classmethod
+    def __get_validators__(cls) -> Iterator[classmethod]:
+        """Gets all validators"""
+        yield cls.validate  # type: ignore
 
 
 @lru_cache()